refactor: update benchmark for nope-validator
